combine nd plugin into core io
