+++ src/ophyd_async/epics/adcore/__init__.py
@@ -1,24 +1,37 @@
-from ._core_io import ADBaseIO, DetectorState, NDFileHDFIO
-from ._core_logic import (DEFAULT_GOOD_STATES, ADBaseShapeProvider,
-                          set_exposure_time_and_acquire_period_if_supplied,
-                          start_acquiring_driver_and_ensure_status)
+from ._core_io import (
+                       ADBaseDataType,
+                       ADBaseIO,
+                       DetectorState,
+                       NDFileHDFIO,
+                       NDPluginStatsIO,
+)
+from ._core_logic import (
+                       DEFAULT_GOOD_STATES,
+                       ADBaseShapeProvider,
+                       set_exposure_time_and_acquire_period_if_supplied,
+                       start_acquiring_driver_and_ensure_status,
+)
 from ._hdf_writer import ADHDFWriter
-from ._nd_plugin import ADBaseDataType, NDPluginStatsIO
 from ._single_trigger import SingleTriggerDetector
-from ._utils import (FileWriteMode, ImageMode, NDAttributeDataType,
-                     NDAttributesXML, stop_busy_record)
+from ._utils import (
+                       FileWriteMode,
+                       ImageMode,
+                       NDAttributeDataType,
+                       NDAttributesXML,
+                       stop_busy_record,
+)
 
 __all__ = [
+    "ADBaseDataType",
     "ADBaseIO",
     "DetectorState",
     "NDFileHDFIO",
+    "NDPluginStatsIO",
     "DEFAULT_GOOD_STATES",
     "ADBaseShapeProvider",
     "set_exposure_time_and_acquire_period_if_supplied",
     "start_acquiring_driver_and_ensure_status",
     "ADHDFWriter",
-    "ADBaseDataType",
-    "NDPluginStatsIO",
     "SingleTriggerDetector",
     "FileWriteMode",
     "ImageMode",
